+++ packages/orq-rc/RELEASES.md
@@ -5369,3 +5369,13 @@ Based on:
 - [typescript v4.1.0-rc.2] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.2 - packages/orq-rc
+
+## 2025-11-25 09:23:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.660.0 (2.760.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.3] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.3 - packages/orq-rc